+++ webeoc/src/main.ts
@@ -1,7 +1,6 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { MicroserviceOptions, Transport } from '@nestjs/microservices';
-import { ComplaintsPublisherService } from './complaints-publisher/complaints-publisher.service';
 import * as dotenv from 'dotenv';
 
 async function bootstrap() {
@@ -15,10 +14,6 @@ async function bootstrap() {
       },
     },
   );
-
-  // Manually trigger the complaints processing
-  const tasksService = app.get(ComplaintsPublisherService);
-  tasksService.handleCron();
 
   await app.listen();
 }
