+++ webeoc/src/webeoc-complaints-scheduler/webeoc-complaints-scheduler.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { WebEOCComplaintsScheduler } from './webeoc-complaints-scheduler.service';
+
+describe('WebEOCComplaintsScheduler', () => {
+  let service: WebEOCComplaintsScheduler;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [WebEOCComplaintsScheduler],
+    }).compile();
+
+    service = module.get<WebEOCComplaintsScheduler>(WebEOCComplaintsScheduler);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
